oldcode/test_key_base.py
import unittest
import os
from PandorasLock.key_base import KeyBase

# Subclass KeyBase for testing, assuming KeyBase is abstract and cannot be instantiated directly
class TestKey(KeyBase):
    # Implement or override any abstract methods or properties if necessary
    pattern_key = "TEST_PATTERN_KEY"

class TestKeyBase(unittest.TestCase):
    def setUp(self):
        # Set up code here. Example: set a path to a test configuration file
        self.config_path = os.path.join(os.path.dirname(__file__), 'test_pandorasconfig.json')
        # Use TestKey for instantiation if KeyBase is abstract
        self.key_base = TestKey(config_path=self.config_path)

    def test_load_regex_patterns(self):
        # Ensure regex patterns are loaded correctly from the config file
        self.assertTrue(isinstance(self.key_base.patterns, dict), "Patterns should be a dictionary")
        self.assertGreater(len(self.key_base.patterns), 0, "Patterns dictionary should not be empty")

    def test_pattern_key_defined(self):
        # Test that the 'pattern_key' is defined in TestKey
        self.assertNotEqual(self.key_base.pattern_key, "", "pattern_key should be defined in TestKey")

# Add more tests as needed to cover other methods and functionalities of KeyBase

if __name__ == '__main__':
    unittest.main()
